+++ app/javascript/components/game/map/MoveTrackOverlay.tsx
@@ -1,15 +1,79 @@
-import React from "react";
+import React, { useState } from "react";
 import Map from "../../../engine/Map";
 import { Coordinate } from "../../../utilities/commonTypes";
 import Hex from "../../../engine/Hex";
-import { circlePath } from "../../../utilities/graphics";
+import { circlePath, roundedRectangle, yMapOffset } from "../../../utilities/graphics";
 import { stateType } from "../../../engine/control/state/BaseState";
+import { moveActions, routActions } from "../../../engine/actions/BaseAction";
+import { executeContextAction, translateAction } from "../../utilities/context";
+import actionsAvailable from "../../../engine/control/actionsAvailable";
+import Game from "../../../engine/Game";
 
 interface MoveTrackOverlayProps {
   map: Map;
+  updateCallback: () => void;
+  scale: number;
+  mapScale: number;
+  svgRef: React.MutableRefObject<HTMLElement>;
 }
 
-export default function MoveTrackOverlay({ map }: MoveTrackOverlayProps) {
+export default function MoveTrackOverlay({
+  map, updateCallback, scale, mapScale, svgRef }: MoveTrackOverlayProps
+) {
+  const [contextMenu, setContextMenu] = useState<JSX.Element | undefined>()
+
+  const contextAction = (type: string) => {
+    if (!map.game) { return }
+    executeContextAction(map.game, undefined, type, updateCallback)
+    setContextMenu(undefined)
+    updateCallback()
+  }
+
+  const rightClick = (event: React.MouseEvent) => {
+    event.preventDefault()
+    if (!map.game) { return }
+    const options = actionsAvailable(map.game, map.game.currentUser, false)
+    const x = (event.clientX - svgRef.current.getBoundingClientRect().x)/scale/mapScale - 24
+    const y = (event.clientY - svgRef.current.getBoundingClientRect().y)/scale/mapScale -
+      yMapOffset/mapScale - 24 + (50 - 50/scale)/mapScale
+    const filtered = options.filter(a => {
+      return ![
+        "sync", "wait", "none", "undo", "join", "leave", "start", "kick", "deploy", "help",
+        "rally_pass", "unselect", "pass", "pass_cancel", "precip_check", "reaction_pass",
+        "initiative", "weather_check", "fire_start_check",
+      ].includes(a.type)
+    }).map(a => a.type)
+    const width = filtered.reduce((max, o) => {
+      const length = translateAction(map.game as Game, undefined, o).length
+      return max > length ? max : length
+    }, 0) * 14.4 + 40
+    const buttons = filtered.map((o, i) => {
+      const text = translateAction(map.game as Game, undefined, o)
+      return (
+        <g key={`context-${i}`} >
+          <path d={roundedRectangle(x + 8, y + 8 + i*36, width, 32, 5)} style={{ fill: "#EEE" }}
+                onClick={() => contextAction(o)}
+                onContextMenu={e => e.preventDefault()} />
+          <text textAnchor="start" x={x + 35} y={y + 30 + i*36} fontSize={24}
+                fontFamily="'Courier Prime', monospace"
+                style={{ fill: "000"}}
+                onClick={() => contextAction(o)}
+                onContextMenu={e => e.preventDefault()} >
+            {text}
+          </text>
+        </g>
+      )
+    })
+    const height = filtered.length * 36 + 12
+    setContextMenu(
+      <g onMouseLeave={() => setContextMenu(undefined)} >
+        <path d={roundedRectangle(x, y, width + 16, height)} style={{ fill: "rgba(0,0,0,0.2)" }} />
+        { buttons }
+      </g>
+    )
+    updateCallback()
+  }
+
   const hexes = (): Hex[] => {
     if (map.game?.gameState?.type === stateType.Move) {
       return map.game.moveState.path.map(p => map.hexAt(new Coordinate(p.x, p.y)) as Hex)
@@ -19,8 +83,7 @@ export default function MoveTrackOverlay({ map }: MoveTrackOverlayProps) {
       return map.game.assaultState.path.map(p => map.hexAt(new Coordinate(p.x, p.y)) as Hex)
     }
     const lastSigAction = map.game?.lastSignificantAction
-    if (lastSigAction &&
-        ["move", "rush", "assault_move", "rout_move", "rout_self"].
+    if (lastSigAction && moveActions.
           includes(lastSigAction.data.action) && lastSigAction.data.path) {
       return lastSigAction.data.path.map(p => map.hexAt(new Coordinate(p.x, p.y)) as Hex)
     }
@@ -31,7 +94,7 @@ export default function MoveTrackOverlay({ map }: MoveTrackOverlayProps) {
   const hexCenters = () => {
     const action = map.game?.lastSignificantAction?.data.action ?? ""
     // TODO: if last sig action was fire, check for displace here
-    const routing = ["rout_move", "rout_self"].includes(action)
+    const routing = routActions.includes(action)
     let first = true
     return hexes().map((h, i) => {
       const offset = Math.max(map.counterDataAt(h.coord).length * 5 - 5, 0)
@@ -39,8 +102,14 @@ export default function MoveTrackOverlay({ map }: MoveTrackOverlayProps) {
       const y = h.yOffset - offset
       const fill = first ? "#AAA" : "#DDD"
       if (!routing) { first = false }
-      return <path key={`${i}-c`} d={circlePath(new Coordinate(x, y), 12)}
-                   style={{ fill, stroke: "#777", strokeWidth: 4 }} />
+      if (routing) {
+        return <path key={`${i}-c`} d={circlePath(new Coordinate(x, y), 12)}
+                    style={{ fill, stroke: "#777", strokeWidth: 4 }} />
+      } else {
+        return <path key={`${i}-c`} d={circlePath(new Coordinate(x, y), 12)}
+                    style={{ fill, stroke: "#777", strokeWidth: 4 }}
+                    onContextMenu={e => rightClick(e)}/>
+      }
     })
   }
 
@@ -72,6 +141,7 @@ export default function MoveTrackOverlay({ map }: MoveTrackOverlayProps) {
       <g>
         { hexTracks() }
         { hexCenters() }
+        { contextMenu }
       </g>
     )
   }