+++ app/javascript/engine/actions/BaseAction.ts
@@ -11,6 +11,7 @@ export const significantActions = [
 ]
 
 export const moveActions = [ "move", "rush", "assault_move", "rout_move", "rout_self" ]
+export const routActions = [ "rout_move", "rout_self" ]
 
 export const fireActions = [ "fire", "intensive_fire", "reaction_fire", "reaction_intensive_fire" ]